requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
cloudinary==1.31.0
DateTime==5.0
Django==4.1.5
django-browser-reload==1.6.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
idna==3.4
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
zope.interface==5.5.2
